Move checkReactivateEligible util
