+++ app/projects/[slug]/project-display.tsx
@@ -8,7 +8,7 @@ import { SignInButton } from '@/components/sign-in-button'
 import { BidAndProfile, BidAndProject } from '@/db/bid'
 import { CommentAndProfile } from '@/db/comment'
 import { Profile } from '@/db/profile'
-import { FullProject, Project } from '@/db/project'
+import { FullProject } from '@/db/project'
 import { Cause, MiniCause } from '@/db/cause'
 import { TxnAndProfiles, TxnAndProject } from '@/db/txn'
 import {
@@ -32,10 +32,6 @@ import { AssuranceBuyBox } from './assurance-buy-box'
 import { calculateTradePoints } from '@/utils/amm'
 import { CertValuationChart } from './valuation-chart'
 import { RichContent } from '@/components/editor'
-import { ReactivateButton, checkReactivateEligible } from './reactivate-button'
-import { buttonClass } from '@/components/button'
-import Link from 'next/link'
-import clsx from 'clsx'
 import { CreatorActionPanel } from './creator-action-panel'
 
 export function ProjectDisplay(props: {
